+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+#### `FlatList`
+
+- Generic data structure.
+- `len` method.
+- `empty` constant for creating empty lists.
+
+#### Tests
+
+- Harness for testing `FlatList`s containing different types..
+- Types for testing with the harness.
+- Fixtures for those types.
+
 ## [0.0.0] - 2025-11-02
 
 - Setup project